Refactor: Move animated components to separate file and optimize
Extracts animated components into a new file and applies optimizations like lazy loading and React.memo.
